Fixed code according to mentor's recommendation

+++ c/grains/src/grains.c
@@ -1,7 +1,7 @@
 #include "grains.h"
 
 uint64_t square(uint8_t index) {
-    return (0 < index && index <= 64) ? 1l << (index - 1) : 0;
+    return (0 < index && index <= 64) ? (uint64_t) 1 << (index - 1) : 0;
 }
 
 uint64_t total(void) {
